feat(simulation): enhance simulation retrieval with detailed results and user authorization

+++ src/main/java/com/renewsim/backend/simulation_service/application/result/SimulationHistoryResultDTO.java
@@ -8,6 +8,6 @@ public record SimulationHistoryResultDTO(
         String energyType,
         double energyGenerated,
         double estimatedSavings,
-        Integer roiYears,
+        Double roiYears,
         LocalDateTime createdAt
 ) {}
